create new libcnet-main.a library to prevent linking error in stdlib tests
