+++ libcnet/utils.h
@@ -66,5 +66,7 @@ void *mem_alloc(int size);
 
 void cnet_free(void *s);
 
+cnet_array *cnet_init_array(int sizei_t, int type_t, int len, int arr_lit_len, ...);
+
 int user_main(cnet_array *args);
 #endif
