+++ tests/stdlib/Makefile
@@ -10,7 +10,7 @@ OBJS := libcnet.a
 OBJS := $(addprefix $(LIB_DIR)/, $(OBJS));
 DEPENDS := string.h utils.h io.h
 DEPENDS := $(addprefix $(LIB_DIR)/, $(DEPENDS));
-CTESTS := cnet_test1 cnet_test2 cnet_test3 cnet_test4 cnet_test5 cnet_test6 cnet_test7 cnet_test8 cnet_test9
+CTESTS := cnet_test1 cnet_test2 cnet_test3 cnet_test4 cnet_test5 cnet_test6 cnet_test7 cnet_test8 cnet_test9 cnet_test10
 
 stdlib_tests: clean stdlib $(CTESTS)
